+++ EnerBank.Interfaces/IResultEstrazione.cs
@@ -4,7 +4,7 @@ namespace EnerBank.Interfaces
 	public interface IResultEstrazione
 	{
 		decimal ImportoTotale { get; set; }
-		ulong TransazioniTotale { get; set; }
+		int TransazioniTotale { get; set; }
 
 	}
 }
